while using this s/w if you face network related then plz refer this http://www.youtube.com/watch?v=fCHUERKhZi8
That link will provide help regarding that problem and full detail how
to solve it.When you will insatll this s/w you must have .net 3.5
framework insatlled and sql server express. If it is not installed in
your system then no need of worry the app will automatically install it
from and no charges for that.